requirements.txt
python-dotenv
langchain
langchain-google-genai
google-generativeai
IPython
chromadb
PyPDF2
python-docx
filetype
docx2txt
langchain-community
langchain-google-vertexai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-aiplatform
matplotlib
seaborn
requests
pandas
openpyxl
xlsxwriter
yfinance
db-dtypes
faiss-cpu
flask
flask-cors
markdown
markdown2
bs4
filetype
yfinance
flask-bcrypt 
flask-jwt-extended 
flask-sqlalchemy 
flask-mail
pyjwt
statsmodels
pmdarima
yahoo_fin
markdownify
prettytable
waitress
apscheduler
dnspython
groq
phidata
duckduckgo-search
openai
mistune
pytz
calcom
requests_cache
apscheduler 
python-dateutil
plyer
pandas-ta
win10toast_click
holidays
openai-whisper
ffmpeg-python
assemblyai
alpha_vantage
win10toast_click --no-deps
plyer
win32api